+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+20.8.0
+------
+
+No Changes
+
 * scrooge: replace deprecated ScalaTest 3.0 API with ScalaTest 3.1/3.2 API. ``PHAB_ID=D525860``
 
 * scrooge-core: Change `c.t.scrooge.ThriftStructMetaData` from an abstract class to a trait